More changes from comments on PR #3554:
- Test `full_transactions` property of `PendingTxSubscription`
- Abstract `SubscriptionContainer` to its own file
